Adapting AT to run on 2018 code (#432)
* Minimal changes to run analyser on 2018 nano

* Fixing a few things

* Updating path for 2018 samples

* Do not recalculate PU weight for 2018 for now

* Updating samples, lumi, and retiring nVetoLeptons cut in 0-lepton channel

* correcting samples

* small change in resubmit_failed_jobs.py to work properly in MultiJobPerCluster mode

* Actually setting puWeight to something nonzero for 2018 MC

* 2018 DY inclusive Ph. Space

* 2018 DY inclusive Ph. Space

* Change settings.txt dataYear value back to 2017

* some plotting changes

* update based on the new sample files

* adding Egamma sample

* error in normalisation of WJets, resolved by remaking sample infos

* changing the xsec values back to the ones corresponding to 2017

* Added Single Muon and DeepCSV scale factors

* Delete unnecessary config files

* adding two scripts for Principal component analysis objects creation

* minor changes

* Restoring 2017 plotting configs and moving 2018 configs to separate ones

* Adding MET trigger efficiency for 2018

* Revert change in settings.txt

* Removing PCA files

+++ resubmit_failed_jobs.py
@@ -124,7 +124,7 @@ for subdir, dirs, files in os.walk(args.dir):
                     missingFiles.append(rootfilename)
     dirpaths = subdir.split('/')
     sample=dirpaths[len(dirpaths)-1]
-    if len(dirpaths) > 1 and not 'HelperClasses' in sample and not 'cfg' in sample and not 'aux' in subdir and not 'plugins' in sample and not args.check and args.fromMultiJobPerCluster:
+    if len(dirpaths) > 1 and not 'HelperClasses' in sample and not 'cfg' in sample and not 'aux' in subdir and not 'plugins' in sample and sample and not args.check and args.fromMultiJobPerCluster:
         try :
             os.remove("%s/REjob%s.submit"%(subdir,sample))
         except OSError:
